Use unisgned instead of size_t in llvm::cl::opt

+++ lib/Transforms/Instrumentation/SimpleMemoryCheck.cc
@@ -28,17 +28,17 @@ static llvm::cl::opt<bool>
                   llvm::cl::desc("Print Simple Memory Check statistics"),
                   llvm::cl::init(false));
 
-static llvm::cl::opt<unsigned int> SMCAnalysisThreshold(
+static llvm::cl::opt<unsigned> SMCAnalysisThreshold(
     "smc-check-threshold",
     llvm::cl::desc("Max no. of analyzed memory instructions"),
     llvm::cl::init(100));
 
-static llvm::cl::opt<size_t> CheckToInstrumentID(
+static llvm::cl::opt<unsigned> CheckToInstrumentID(
     "smc-instrument-check",
     llvm::cl::desc("Id of the check to instrument"),
     llvm::cl::init(0));
 
-static llvm::cl::opt<size_t> AllocToInstrumentID(
+static llvm::cl::opt<unsigned> AllocToInstrumentID(
     "smc-instrument-alloc",
     llvm::cl::desc("Id of the allocation site to instrument"),
     llvm::cl::init(0));
